requirements/base.txt
# Base/shared runtime dependencies (no dev/test tools)

# Core Telegram Bot
python-telegram-bot[job-queue]==22.5

# Build tooling (needed for wheels during installs)
setuptools>=78.1.1,<81

# Database
pymongo==4.15.3
motor==3.7.1

# Code Processing & Language Detection
pygments==2.19.2
python-magic==0.4.27
chardet==5.2.0
langdetect==1.0.9
textstat==0.7.10

# Web Requests & APIs
requests==2.32.5
aiohttp==3.13.1
httpx==0.28.1

# Google APIs (Drive OAuth + Client)
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2

# GitHub Integration
PyGithub==2.8.1

# File Processing & Env
python-dotenv==1.2.1
# pathlib2 הוסר: לא נחוץ ב-Python 3

# Image Processing (for syntax highlighting images)
Pillow==12.0.0
cairosvg==2.8.2

# High-quality HTML→Image rendering (Chromium via Playwright)
playwright==1.49.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
click==8.3.0
arrow==1.4.0

# Logging & Monitoring
structlog==25.4.0
sentry-sdk[flask]==2.42.1
psutil==7.1.2
prometheus-client==0.23.1

# OpenTelemetry (tracing & basic auto-instrumentation)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-pymongo

# Security
bcrypt==5.0.0
cryptography==46.0.3

# Web Server & HTTP layer
Flask==3.1.2
gunicorn==23.0.0
uvicorn==0.38.0
whitenoise==6.11.0
Flask-Compress==1.20

# Rate limiting (Flask + generic)
Flask-Limiter==4.0.0
limits==5.6.0


# Async Support
aiofiles==25.1.0

# Cache
redis==7.0.0

# Configuration
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0

# HTML/Text Processing
beautifulsoup4==4.14.2
markdown==3.9
html2text==2025.4.15

# Version Control Integration
gitpython==3.1.45

# API Rate Limiting
ratelimit==2.2.1
slowapi==0.1.9

# Text Processing
rapidfuzz==3.14.1

# CLI Tools
# Flask-Limiter 3.5.0 דורש rich<14 — נצמיד לגרסה יציבה תואמת
rich==14.2.0
typer==0.20.0

# Resilience utilities
tenacity==9.1.2
